requirements.txt
Pillow>=8.2
freetype-py>=2.2